Terminology fixes
Use "regular PostgreSQL" instead of "vanilla PostgreSQL". We currently use both. We need to settle on one. I think "regular" is less ambiguous than "vanilla" for users who are not familiar with PostgreSQL lingo.

I will add a glossary definition for "regular PostgreSQL" in a separate PR for the glossary.