Update node.md to make it consistent with the Neon console
The examples given in the file were not consistent with the output gotten from the Neon console: For nodes.js the console is not giving a connection string but individual variables. The example program for postgres.js in the console is using those variables. Now both examples in this document are using them as well to be consistent with the console.

In addition removed the usage of the endpoint id from the example. It is no longer needed for pg and postgres.js.

Also made the two examples a bit more consistent.